moved funtions to image object matching module
